Changes improving IPv6 and adding SRv6 encap/decap and some processing
capabilities to showcase the HIRT paper to be presented at ICNP'24.

It also adds the library elements fo network-layer FEC used in the
aforementioned paper. However the FEC elements themselves use RLC code
that is under patent, so if you want them please ask Louis. For the same
reasons, I had to squash all commits as once, sorry about the huge
change.

Improvements:
 - updated gitlab CI with much more tests
 - SRV6 encap and decap elements
 - Fix compatibility problem with aarch64
 - Fix compatibility issues with MacOS

@louisna

 was mostly involved in the code, and then me.

i

+++ elements/standard/meter.cc
@@ -29,12 +29,12 @@ Meter::push(int, Packet *p)
 {
   _rate.update(1);		// packets, not bytes
 
-  uint64_t r = _rate.scaled_average();
+  unsigned r = _rate.scaled_average();
   if (_nmeters < 2) {
     int n = (r >= _meter1);
     output(n).push(p);
   } else {
-    uint64_t *meters = _meters;
+    unsigned *meters = _meters;
     int nmeters = _nmeters;
     for (int i = 0; i < nmeters; i++)
       if (r < meters[i]) {
