+++ elements/standard/shaper.cc
@@ -32,7 +32,7 @@ Shaper::Shaper()
 int
 Shaper::configure(Vector<String> &conf, ErrorHandler *errh)
 {
-    uint64_t rate;
+    uint32_t rate;
     Args args(conf, this, errh);
     if (is_bandwidth())
 	args.read_mp("RATE", BandwidthArg(), rate);
