+++ include/click/bigint.hh
@@ -384,8 +384,8 @@ class Bigint { public:
 
 };
 
-/** @brief Typical Bigint usage with uint64_t limb_type. */
-typedef Bigint<uint64_t> bigint;
+/** @brief Typical Bigint usage with uint32_t limb_type. */
+typedef Bigint<uint32_t> bigint;
 
 CLICK_ENDDECLS
 #endif
